Break App.js into Components
