+++ src/containers/mainmenu.js
@@ -1,0 +1,11 @@
+import React, {Component} from 'react'
+
+class MainMenu extends Component {
+	render() {
+		return (
+			<button className="button play green" disabled="true">Play</button>
+		)
+	}
+}
+
+export default MainMenu
