+++ src/containers/scoreboard.js
@@ -1,0 +1,11 @@
+import React, {Component} from 'react'
+
+class Scoreboard extends Component {
+	render() {
+		return (
+			<div>Score: 23</div>
+		)
+	}
+}
+
+export default Scoreboard
