+++ src/containers/timetracker.js
@@ -1,0 +1,11 @@
+import React, {Component} from 'react'
+
+class TimeTracker extends Component {
+	render() {
+		return (
+			<div>Time Left: 157</div>
+		)
+	}
+}
+
+export default TimeTracker
