+++ src/index.js
@@ -1,6 +1,8 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
-import App from './App';
+import {Provider} from 'react-redux'
+import {createStore} from 'redux'
+import App from './components/App';
 import './index.css';
 
 ReactDOM.render(
